==> test_scripts/analysis/02_generate_plots.R <==
library(readr)
library(jsonlite)
library(data.table)
library(tidyverse)
library(ggplot2)
library(rjson)
library(rlist)
library(future)
library(furrr)
library(parallel)
library(foreach)
library(doParallel)
library(tictoc)
library(plotly)
require(magick)
require(cowplot)
source("00_functions.R")
source("00_viz_functions.R")

tic("picture generation took")
print("==============================================================")
back_up_wd <- getwd()
current_dir <- system("pwd",intern = TRUE)
current_dir <- "/Users/nikitakricko/Documents/GitHub/2021_10_arm_cpu_comparison_m6/test_scripts/analysis"
setwd(current_dir)
path_to_save_plots <- paste(current_dir, "/autogen_plots", sep="")
dir.create(path_to_save_plots)
setwd(path_to_save_plots)
print(current_dir)
# current_dir <- "/Users/nikitakricko/Documents/GitHub/2021_10_arm_cpu_comparison_c5/test_scripts/analysis"
print("==============================================================")
path_to_file <- paste(current_dir,"oltp_sysbench_logs.fst", sep="/")

oltp_test_result <- fst::read.fst(path_to_file) %>% as.data.table()


condition_list <- c("m6i", "m6g", "m6a")
oltp_test_result <- oltp_test_result[VM_type %like_in% condition_list]
# oltp_test_result <- oltp_test_result %>% like_in("VM_type",condition_list)


col_list <- c("read" ,"query_exec_other",
              "query_exec_general_statistic_total_number_of_events","latency_ms_min", "latency_ms_avg", "latency_ms_max",
              "latency_ms_95th_percentile", "latency_ms_sum", "latency_ms_events_avg", "latency_ms_events_stddev",
              "latency_ms_execution_time_avg","latency_ms_execution_time_stddev", "sql_stat_transactions_total",
              "sql_stat_transactions_per_sec","queries_total", "queries_per_sec" )



cpuList <- oltp_test_result$cpu_amount %>% unique()
ec2_types <- oltp_test_result$ec2_type %>% unique()




for(column in col_list){

  print(column)
  autogenerated_title <- paste(column,"autogenerated", sep = "_")
  print(autogenerated_title)
  auto_plot <- get_db_universal(oltp_test_result,
                           input_title = autogenerated_title,
                           column_name = column,
                           yaxis_label = column,
                           input_subtitle = "",
                           input_caption = "sysbench: 1.0.18, MySQL: 8.0.26-0",
                           x_axis_aws = FALSE,
                           add_logo = TRUE,
                           facet_cpu=TRUE)
  save_plot(autogenerated_title, auto_plot)

  for(cpu_num in cpuList){
    autogenerated_title <- paste(column, "for", cpu_num,"cpu","autogenerated", sep = "_")
    print(autogenerated_title)
    auto_plot <- get_db_universal(oltp_test_result[cpu_amount == cpu_num],
                                  input_title = autogenerated_title,
                                  column_name = column,
                                  yaxis_label = column,
                                  input_subtitle = "",
                                  input_caption = "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                  x_axis_aws = FALSE,
                                  add_logo = TRUE,
                                  facet_cpu=TRUE)
    save_plot(autogenerated_title, auto_plot)
  }

  for(ec2 in ec2_types){
    autogenerated_title <- paste(column, "for",ec2,"autogenerated", sep = "_")
    print(autogenerated_title)
    auto_plot <- get_db_universal(oltp_test_result[ec2_type == ec2],
                                  input_title = autogenerated_title,
                                  column_name = column,
                                  yaxis_label = column,
                                  input_subtitle = "",
                                  input_caption = "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                  x_axis_aws = FALSE,
                                  add_logo = TRUE,
                                  facet_cpu=TRUE)
    save_plot(autogenerated_title, auto_plot)
  }
}










### Clean plots ################################
### small
table(oltp_test_result$VM_type)


p_11 <- get_db_universal(oltp_test_result[ec2_type == "small"],
                         input_title = "Workload: queries per second (vCPU: 2, 4,8)",
                         column_name = "queries_per_sec", 
                         yaxis_label = "Amount of queries per second",
                         x_axis_aws = FALSE,
                         add_logo = FALSE,
                         facet_cpu=TRUE)
save_plot("11_small_workload_qps.png", p_11)

p_12 <- get_db_universal(oltp_test_result[ec2_type == "small"],
                         input_title = "p95 latency during test (vCPU: 2, 4,8)",
                         column_name = "latency_ms_95th_percentile", 
                         yaxis_label = "ms",
                         x_axis_aws = FALSE,
                         add_logo = TRUE,
                         facet_cpu=TRUE)
save_plot("12_small_95th_latency.png", p_12)
### medium
p_21 <- get_db_universal(oltp_test_result[ec2_type != "small"],
                         input_title = "Workload: queries per second (vCPU: 16,48,64)",
                         column_name = "queries_per_sec", 
                         yaxis_label = "Amount of queries per second",
                         x_axis_aws = FALSE,
                         add_logo = TRUE,
                         facet_cpu=TRUE)
save_plot("21_ML_workload_qps.png", p_21)

p_22 <- get_db_universal(oltp_test_result[ec2_type != "small"],
                         input_title = "p95 latency during test  (vCPU: 16,48,64)",
                         column_name = "latency_ms_95th_percentile", 
                         yaxis_label = "ms",
                         x_axis_aws = FALSE,
                         add_logo = TRUE,
                         facet_cpu=TRUE)
save_plot("22_ML_95th_latency.png", p_22)

# ### large
# p_31 <- get_db_universal(oltp_test_result[ec2_type == "large"],
#                          input_title = "Workload: queries per second (vCPU: 48,64)",
#                          column_name = "queries_per_sec", 
#                          yaxis_label = "Amount of queries per second",
#                          x_axis_aws = FALSE,
#                          add_logo = FALSE,
#                          facet_cpu=TRUE)
# save_plot("03_rel_com_overview_intel.png", p_31)
# 
# p_32 <- get_db_universal(oltp_test_result[ec2_type == "large"],
#                          input_title = "p95 latency during test  (vCPU: 48,64)",
#                          column_name = "latency_ms_95th_percentile", 
#                          yaxis_label = "ms",
#                          x_axis_aws = FALSE,
#                          add_logo = FALSE,
#                          facet_cpu=TRUE)
# save_plot("03_rel_com_overview_intel.png", p_32)
### overview
p_41 <- get_db_universal(oltp_test_result,
                         input_title = "Workload: queries per second",
                         column_name = "queries_per_sec", 
                         yaxis_label = "Amount of queries per second",
                         x_axis_aws = FALSE,
                         add_logo = TRUE,
                         facet_cpu=TRUE)
save_plot("41_overview_workload_qps.png", p_41)

p_42 <- get_db_universal(oltp_test_result,
                        input_title = "p95 latency during test",
                        column_name = "latency_ms_95th_percentile", 
                        yaxis_label = "ms",
                        x_axis_aws = FALSE,
                        add_logo = TRUE,
                        facet_cpu=TRUE)
save_plot("42_overview_95th_percentile.png", p_42)



### relative comparison #########


rel_com_s <- intel_graviton_comparison(oltp_test_result[ec2_type == "small"],
                                       input_subtitle="",
                                       input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                       input_relative=TRUE)
save_plot("131_IS_relative_comparison.png", rel_com_s)

rel_com_ML <- intel_graviton_comparison(oltp_test_result[ec2_type != "small"],
                                       input_subtitle="",
                                       input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                       input_relative=TRUE)
save_plot("231_IML_relative_comparison.png", rel_com_ML)

# rel_com_f <- intel_graviton_comparison(oltp_test_result[ec2_type == "large"],
#                                        input_subtitle="",
#                                        input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
#                                        input_relative=TRUE)
# save_plot("03_IL_relative_comparison.png", rel_com_f)

rel_com_AS <- intel_graviton_comparison(oltp_test_result[ec2_type == "small"],
                                       input_subtitle="",
                                       input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                       input_relative=TRUE,
                                       target_comparing_cpu="AMD")
save_plot("132_AS_relative_comparison.png", rel_com_AS)

rel_com_AML <- intel_graviton_comparison(oltp_test_result[ec2_type != "small"],
                                       input_subtitle="",
                                       input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                       input_relative=TRUE,
                                       target_comparing_cpu="AMD")
save_plot("232_AML_relative_comparison.png", rel_com_AML)

# rel_com_AL <- intel_graviton_comparison(oltp_test_result[ec2_type == "large"],
#                                        input_subtitle="",
#                                        input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
#                                        input_relative=TRUE,
#                                        target_comparing_cpu="amd")
# save_plot("03_AL_relative_comparison.png", rel_com_AL)

rel_com_overview_intel <- intel_graviton_comparison(oltp_test_result,
                                                    input_subtitle="",
                                                    input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                                    input_relative=TRUE,
                                                    target_comparing_cpu="intel")
save_plot("431_rel_com_overview_intel.png", rel_com_overview_intel)

rel_com_overview_amd <- intel_graviton_comparison(oltp_test_result,
                                                  input_subtitle="",
                                                  input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                                  input_relative=TRUE,
                                                  target_comparing_cpu="amd")
save_plot("432_rel_com_overview_AMD.png", rel_com_overview_amd)




### absolute comparison ##########


abs_com_s <- intel_graviton_comparison(oltp_test_result[ec2_type == "small"],
                                       input_subtitle="",
                                       input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                       input_relative=FALSE)
save_plot("141_abs_com_s.png", abs_com_s)

abs_com_ml <- intel_graviton_comparison(oltp_test_result[ec2_type != "small"],
                                       input_subtitle="",
                                       input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                       input_relative=FALSE)
save_plot("241_abs_com_ml.png", abs_com_ml)

# abs_com_f <- intel_graviton_comparison(oltp_test_result[ec2_type == "large"],
#                                        input_subtitle="",
#                                        input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
#                                        input_relative=FALSE)
# save_plot("04_abs_com_f.png", abs_com_f)


abs_com_AS <- intel_graviton_comparison(oltp_test_result[ec2_type == "small"],
                                       input_subtitle="",
                                       input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                       input_relative=FALSE,
                                       target_comparing_cpu="AMD")
save_plot("142_abs_com_s.png", abs_com_AS)

abs_com_AML <- intel_graviton_comparison(oltp_test_result[ec2_type != "small"],
                                       input_subtitle="",
                                       input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                       input_relative=FALSE,
                                       target_comparing_cpu="AMD")
save_plot("242_abs_com_ml.png", abs_com_AML)

# abs_com_AL <- intel_graviton_comparison(oltp_test_result[ec2_type == "large"],
#                                        input_subtitle="",
#                                        input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
#                                        input_relative=FALSE,
#                                        target_comparing_cpu="amd")
# save_plot("04_abs_com_f.png", abs_com_AL)
abs_com_overview_intel <- intel_graviton_comparison(oltp_test_result,
                                                    input_subtitle="",
                                                    input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                                    input_relative=FALSE)
save_plot("441_abs_com_overview_intel.png", abs_com_overview_intel)

abs_com_overview_amd <- intel_graviton_comparison(oltp_test_result,
                                                  input_subtitle="",
                                                  input_caption= "sysbench: 1.0.18, MySQL: 8.0.26-0",
                                                  input_relative=FALSE,
                                                  target_comparing_cpu="AMD")
save_plot("442_abs_com_overview_amd.png", abs_com_overview_amd)


### requests per_dollar ###########
p_05_rpd_overview <- requests_per_dollar(oltp_test_result)
save_plot("05_p_05_rpd_overview.png", p_05_rpd_overview)
p_05_rpd_s <- requests_per_dollar(oltp_test_result[ec2_type == "small"])
save_plot("05_p_05_rpd_small.png", p_05_rpd_s)
p_05_rpd_f <- requests_per_dollar(oltp_test_result[ec2_type != "small"])
save_plot("05_p_05_rpd_large.png", p_05_rpd_f)

### request per hour  ##############
p_06_rph_overview <- requests_ph(oltp_test_result)
save_plot("06_p_06_rph_overview.png", p_06_rph_overview)
p_06_rph_s <- requests_ph(oltp_test_result[ec2_type == "small"])
save_plot("06_p_06_rph_small.png", p_06_rph_s)
p_06_rph_f <- requests_ph(oltp_test_result[ec2_type != "small"])
save_plot("06_p_06_rph_large.png", p_06_rph_f)

#### eficient
p_07_efficiency_overview <- efficient_comparison_point_plot(oltp_test_result, facet_threads=FALSE)
# print(p_10_overview)
save_plot("07_p_07_efficiency_overview.png", p_07_efficiency_overview)
  
toc()
